refactor(components-name): Updates components name to capitalize the first letter.
