+++ src/app/layout.tsx
@@ -14,7 +14,7 @@ const geistMono = Geist_Mono({
 
 export const metadata: Metadata = {
   title: "Readowl - Cultivando literatura em qualquer lugar",
-  description: "Generated by create next app",
+  description: "Readowl é uma plataforma gratuita para publicação e leitura de livros brasileiros. Descubra novos autores, publique suas próprias obras e faça parte de uma comunidade apaixonada por literatura."
 };
 
 export default function RootLayout({
@@ -23,7 +23,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
+    <html lang="pt-br">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
