+++ src/components/sections/HowItHelps.tsx
