+++ src/components/sections/HowToPost.tsx
