+++ src/components/sections/Purpose.tsx
